make timeLabel to show timer

+++ Chew Play/CaptureManager/CaptureManager.swift
@@ -12,7 +12,7 @@ protocol CaptureManager {
     var store: Store { get }
     var previewLayer: CALayer? { get }
     var isChewing: AnyPublisher<Bool, Never> { get }
-    var progress: AnyPublisher<Float, Never> { get }
+    var progress: AnyPublisher<Int, Never> { get }
     func setup()
     func setOnSetPreviewLayer(_ value: ((CALayer) -> Void)?)
 }
